requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
webdriver-manager>=4.0.0
pandas>=2.1.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
rich>=13.7.0
flask>=3.0.0
gunicorn>=21.2.0
